fix: compute mask_level thresholds in f32, as C does
SE and PE are now byte-identical to bwa-mem2 on 400k real HG002 reads/pairs.

mask_level and mask_level_redun are C `float`s, so `or_ > opt->mask_level_redun * mr`
converts the int64 to float and compares in single precision. We widened to f64,
which changes the answer exactly at the boundary:

  0.95f is really 0.949999988079071044921875
  f32: 0.95f * 40 rounds back to exactly 38.0  -> 38 > 38.0  is false -> hit kept
  f64: 0.95f * 40 stays  37.99999952316284     -> 38 > that  is true  -> hit dropped

That is one region's fate per boundary case, and it cascaded: mem_mark_primary_se
sets hash = hash_64(id + i) from each region's *index*, so dropping a region
shifts every later hash and re-orders the alnreg_hlt tie-break, changing which of
two equal-scoring hits is emitted. Diffing the pre-sort region arrays against the
instrumented oracle showed 321 regions against bwa's 324, the first 28 identical.

Fixed in mem_sort_dedup_patch's redundancy test, mem_mark_primary_se_core's
overlap test, and mem_pair's cal_sub. Also fixed mem_reg2sam's drop_ratio test,
which truncated the product to i32 before comparing instead of comparing in float.

Left as-is (verified against the C, which really is double there): mem_gen_alt's
get_pri_idx takes `double XA_drop_ratio`, and MIN_RATIO is a double literal.
bwa-chain's mem_chain_flt was already f32.

Session totals on real HG002 data, full-line diffs vs bwa-mem2:
  SE  666 -> 0   (402562 records, cmp clean)
  PE 2326 -> 0   (805261 records, cmp clean)
Simulated oracle 5000/5000; 68/68 tests.

+++ crates/bwa-mem/src/pe.rs
@@ -546,7 +546,8 @@ fn cal_sub(opt: &MemOpt, r: &[MemAlnReg]) -> i32 {
         let e_min = r[j].qe.min(r[0].qe);
         if e_min > b_max {
             let min_l = (r[j].qe - r[j].qb).min(r[0].qe - r[0].qb);
-            if f64::from(e_min - b_max) >= f64::from(min_l) * f64::from(opt.mask_level) {
+            // f32: C computes `min_l * opt->mask_level` in single precision.
+            if (e_min - b_max) as f32 >= min_l as f32 * opt.mask_level {
                 break;
             }
         }
@@ -1059,7 +1060,7 @@ fn mem_reg2sam(
             continue; // !MEM_F_ALL: drop all secondaries
         }
         if p.secondary >= 0
-            && p.score < (a[p.secondary as usize].score as f32 * opt.drop_ratio) as i32
+            && (p.score as f32) < a[p.secondary as usize].score as f32 * opt.drop_ratio
         {
             continue;
         }
